+++ packages/guideline-blocks-settings/src/components/Link/LinkInput.tsx
@@ -65,9 +65,15 @@ export const LinkInput = ({
                 <div className="tw-mt-3">
                     <LinkSelector
                         url={url}
-                        appBridge={appBridge}
                         onUrlChange={onUrlChange}
                         buttonSize={buttonSize ?? ButtonSize.Medium}
+                        getAllDocuments={() => appBridge.getAllDocuments()}
+                        getDocumentPagesByDocumentId={(documentId) =>
+                            appBridge.getDocumentPagesByDocumentId(documentId)
+                        }
+                        getDocumentSectionsByDocumentPageId={(documentPageId) =>
+                            appBridge.getDocumentSectionsByDocumentPageId(documentPageId)
+                        }
                     />
                 </div>
             )}